Add timestamping and fix buggy logic
On request the server shouldn't call its reply rpc, but should instruct its client to call the reply rpc, so that the correct server receives the reply.

Added reminders for adding queuing logic and an exit method which must call reply to all the clients in the queue.

In the request rpc method, the server just checked if it has a lower lamport clock, when it wants the critical section, but should also compare against the priority of the server and the requestee as per the algorithm on slides